Sentinel: split tests
This allows more control and isolation of the failover portion (as with other tests). Also adds a bit more logging as to endpoint state to help debug issues.
